requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
pandas==1.3.1
Pillow==8.3.1
pyparsing==2.4.7
PySimpleGUI==4.45.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
xlrd==2.0.1
